+++ src/util/unicode.cpp
@@ -36,7 +36,7 @@ std::string narrow(const wchar_t *s)
   int slength=static_cast<int>(wcslen(s));
   int rlength=
     WideCharToMultiByte(CP_UTF8, 0, s, slength, NULL, 0, NULL, NULL);
-  std::string r(rlength, 0);
+  std::string r(narrow_cast<std::size_t>(rlength), 0);
   WideCharToMultiByte(CP_UTF8, 0, s, slength, &r[0], rlength, NULL, NULL);
   return r;
 
@@ -61,7 +61,7 @@ std::wstring widen(const char *s)
   int slength=static_cast<int>(strlen(s));
   int rlength=
     MultiByteToWideChar(CP_UTF8, 0, s, slength, NULL, 0);
-  std::wstring r(rlength, 0);
+  std::wstring r(narrow_cast<std::size_t>(rlength), 0);
   MultiByteToWideChar(CP_UTF8, 0, s, slength, &r[0], rlength);
   return r;
 
@@ -86,7 +86,7 @@ std::string narrow(const std::wstring &s)
   int slength=static_cast<int>(s.size());
   int rlength=
     WideCharToMultiByte(CP_UTF8, 0, &s[0], slength, NULL, 0, NULL, NULL);
-  std::string r(rlength, 0);
+  std::string r(narrow_cast<std::size_t>(rlength), 0);
   WideCharToMultiByte(CP_UTF8, 0, &s[0], slength, &r[0], rlength, NULL, NULL);
   return r;
 
@@ -103,7 +103,7 @@ std::wstring widen(const std::string &s)
   int slength=static_cast<int>(s.size());
   int rlength=
     MultiByteToWideChar(CP_UTF8, 0, &s[0], slength, NULL, 0);
-  std::wstring r(rlength, 0);
+  std::wstring r(narrow_cast<std::size_t>(rlength), 0);
   MultiByteToWideChar(CP_UTF8, 0, &s[0], slength, &r[0], rlength);
   return r;
 
@@ -160,7 +160,7 @@ std::vector<std::string> narrow_argv(int argc, const wchar_t **argv_wide)
     return std::vector<std::string>();
 
   std::vector<std::string> argv_narrow;
-  argv_narrow.reserve(argc);
+  argv_narrow.reserve(narrow_cast<std::size_t>(argc));
 
   for(int i=0; i!=argc; ++i)
     argv_narrow.push_back(narrow(argv_wide[i]));
